[Home] removes config relationship
