+++ src/main/core/Resources/config/suggested/doctrine_prod.yml
@@ -8,6 +8,8 @@ doctrine:
         password:       "%database_password%"
         port:           "%database_port%"
         charset:        UTF8
+        options:
+            1002: 'SET sql_mode=(SELECT REPLACE(@@sql_mode, "ONLY_FULL_GROUP_BY", ""))'
 
     orm:
         auto_generate_proxy_classes: "%kernel.debug%"
